added comunicatiojn with database and a test profile view
I added a class to be used to upload data to the database and a page where the profile will need to be built so we can continue with the next implimentations.
log out works it is up to you to decide where to place it
still need to figure out how to get the images for the profile pic back from the database.